requirements.txt
gradio
Pillow
pytesseract
